hook de setup role
